# docs/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Categoria": {
				"properties": {
					"path": {
						"type": "string"
					},
					"nome": {
						"type": "string"
					}
				},
				"required": [
					"path",
					"nome"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Comentario": {
				"properties": {
					"id": {
						"type": "string"
					},
					"idPai": {
						"type": "string"
					},
					"timestamp": {
						"type": "number",
						"format": "double"
					},
					"corpo": {
						"type": "string"
					},
					"autor": {
						"type": "string"
					},
					"nota": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"idPai",
					"timestamp",
					"corpo",
					"autor",
					"nota"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_Comentario.autor-or-corpo_": {
				"properties": {
					"autor": {
						"type": "string"
					},
					"corpo": {
						"type": "string"
					}
				},
				"required": [
					"autor",
					"corpo"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"ComentarioParams": {
				"$ref": "#/components/schemas/Pick_Comentario.autor-or-corpo_"
			},
			"OpcaoVoto.Positivo": {
				"enum": [
					"positivo"
				],
				"type": "string"
			},
			"OpcaoVoto.Negativo": {
				"enum": [
					"negativo"
				],
				"type": "string"
			},
			"Voto": {
				"properties": {
					"opcao": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/OpcaoVoto.Positivo"
							},
							{
								"$ref": "#/components/schemas/OpcaoVoto.Negativo"
							}
						]
					}
				},
				"required": [
					"opcao"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Post": {
				"properties": {
					"id": {
						"type": "string"
					},
					"titulo": {
						"type": "string"
					},
					"corpo": {
						"type": "string"
					},
					"autor": {
						"type": "string"
					},
					"categoria": {
						"type": "string"
					},
					"timestamp": {
						"type": "number",
						"format": "double"
					},
					"nota": {
						"type": "number",
						"format": "double"
					},
					"numeroComentarios": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"titulo",
					"corpo",
					"autor",
					"categoria",
					"timestamp",
					"nota",
					"numeroComentarios"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_Post.titulo-or-corpo-or-autor-or-categoria_": {
				"properties": {
					"autor": {
						"type": "string"
					},
					"corpo": {
						"type": "string"
					},
					"titulo": {
						"type": "string"
					},
					"categoria": {
						"type": "string"
					}
				},
				"required": [
					"autor",
					"corpo",
					"titulo",
					"categoria"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"PostParams": {
				"$ref": "#/components/schemas/Pick_Post.titulo-or-corpo-or-autor-or-categoria_"
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer"
			}
		}
	},
	"info": {
		"title": "forum-api-client",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Claudio Romero"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/categorias": {
			"get": {
				"operationId": "Listar",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Categoria"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "Criar",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Categoria"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Categoria"
							}
						}
					}
				}
			}
		},
		"/categorias/{id}": {
			"delete": {
				"operationId": "Excluir",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/comentarios/{idPai}": {
			"get": {
				"operationId": "Listar",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Comentario"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "idPai",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "Criar",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Comentario"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "idPai",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ComentarioParams"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "Atualizar",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Comentario"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "idPai",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ComentarioParams"
							}
						}
					}
				}
			}
		},
		"/comentarios/{idComentario}/votar": {
			"put": {
				"operationId": "Votar",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Comentario"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "idComentario",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Voto"
							}
						}
					}
				}
			}
		},
		"/comentarios/{idComentario}": {
			"delete": {
				"operationId": "Excluir",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "idComentario",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/posts": {
			"get": {
				"operationId": "Listar",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Post"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "path",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/posts/{idPost}": {
			"get": {
				"operationId": "ObterPorId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Post"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "idPost",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "Atualizar",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Post"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "idPost",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PostParams"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Excluir",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "idPost",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/posts/{idPost}/votar": {
			"put": {
				"operationId": "Votar",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Post"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "idPost",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Voto"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}